Remove forwarding res to is_allocated on unique_resource_data construction.
It is possible that the resource source will become moved-from after the
call to is_allocated, and that moved-from state will then be used to
initialize the resource in the resource_holder. To avoid this, don't
forward the resource source to is_allocated.

Fixes https://github.com/Lastique/scope/issues/10.