Added a workaround in tests for MSVC 14.1 in C++17 mode.
The compiler ICEs when tests for unallocated_resource helper are
compiled. It isn't clear what upsets the compiler, but changing the
unique_resource constructor that is being called works around the
problem.

Since unallocated_resource is still usable, we're not disabling it
for this compiler and just update the test. Users encountering ICEs
will have to deal with them on their end (upgrading the compiler might
be one way to do it).

https://developercommunity.visualstudio.com/t/ICE-when-compiling-NTTP-related-BoostSc/10922599